Write the system sleep value to system settings
We were saving it to local shared prefs but had no way for another
package to get it as we lost the provider for it. This moves the value
to system settings where it should be.

This also required one string change to match the behavior of the new
setting.

Also cleans up the naming so the code is easier to read.

bug:17665809
Change-Id: Ib2d8860909c946daac16c4ccf40de808c274e9f9