Created always on broadcast receiver for BT accessories
Before we only added/removed from our list of connected BT accessories
when settings was open, meaning if a device was connected when settings
was not opened, it would not be listed as "connected" in settings.

Created a broadcast receiver which is always listening for
connect/disconnect events and stores that device's address to set of
strings in sharedprefs.  This list is cleared on boot.

b/17462341

Change-Id: I7037ed766fcde1c4d9e1b7f47763e1e5c549f3f8