About: Fix greyed out titles on certain settings
Explicitly setting enabled(false) causes an item to be non clickable
and "beep" when clicked, however it also causes the title to be greyed
out, which is inconsistent with the UX of the other items in the menu.
The UX is consistent in this manner, but it does make the item clickable
it just does nothing.

Change-Id: I33e56a76d75b508e904ca081b3ebc81d057de595