Enable using SetupWraith Account Flow
Bug: 34470454

As part of the transition of the google account flow from
NoTouchAuthDelegate to SetupWraith, this CL allows a path
to use the SetupWraith AccountSetupActivity with
MinuteMaid. With the use of system property "useSuwAcctFlow"
we can delegate between the SetupWraith flow, or the
existing flow in NoTouchAuthDelegate. After the GMSCore
changes are made available in master, we can remove the
conditional and just use the flow in SetupWraith.

Change-Id: Id0b635097ee9c388fcd5e897ace952ebed6f3498