Added device tutorials entry to TvSettings
Created "Device Tutorials" entry in the "About" section of Settings.
This preference becomes visible if a system app is found that handles
action: com.google.android.tvtutorials.action.FEATURE_TUTORIALS_TOUR
This action launches TvTutorials app which will display tours for
individual features of the current device.
An icon will be displayed next to the preference if an app icon is
defined for TvTutorials.

Test: Tested with the POST_SETUP_TOUR action that's currently handled by
TvTutorials.
Change-Id: I19efde3577cc0a5931ede3eb9c334ebfb994f458