Set wifi password minimal length to 5
WEP and PSK use the same password field, but the original code
flow only consider the minimal length of PSK. So it will cause
user can not input WEP password which less than 8.
So we need to consider the minimal length of WEP = 5.
Besides, we only allow 5, 13, 16 characters password if type is WEP.

Change-Id: I490145c015c72e2f164834ae4067f1b508865bdd