Change Wifi finish animation for Setup
Bug: 37576742
Test: Manual test through setup

Screenrecord: https://drive.google.com/open?id=0B_p2XYepQoFTSUNBSWVqN1JJX2M

This CL does the following:

1) The animation which runs when the activity finishes is
changed (while being run in setup only) to slide out and fade
instead of stationary fading -- We have to track the result
of the activity in order to decide if we should slide out left
or slide right.

2) Adds an extra to decide from which direction the first fragment
should enter and changes the boolean in dispayPage if this extra
is set only for the first fragment. This corrects the animation
when we go backwards in the setup flow. *requires change to intent
in SetupWraith (see ag/2196635)

Change-Id: I02e304131390019d46f47e757329031b1d2a8955