Fix DeviceName Animations
Bug: 37576742
Bug: 38232038
Test: Manual through setup and settings (about > device name)

Screenrecords available in bug tickets

This CL does the following:

1) Fix overlapping fragment transitions in the device name flow
in settings by preserving the enter transition of the
DeviceNameSetFragment, via super.onProvideFragmentTransitions().
Because we are adding this fragment asRoot in setup,
the enter transition is ignored and instead the animation is
handled in the activity. (see https://developer.android.com/reference/android/support/v17/leanback/app/GuidedStepFragment.html#UI_STYLE_ACTIVITY_ROOT)

2) Fixes entry animation for setup. The activity will start an
animation based on the EXTRA_MOVING_FORWARD that is passed through
the intent. *This requires a small CL in SetupWraith*

3) Fixes close animation for setup -- We are using a boolean (mResultOk)
to track the result of the activity. This is used to determine the
direction of our animation based on which direction we move in the
setup flow. activityCloseExitAnimation was set to null because we
handle this at runtime.

Change-Id: Ibb70fe87ab1a5f089efaa885e88cfddf6d35226f