Replace ActivityCloseExitAnimation in Setup
Bug: 38316198
Test: Manually through setup

The default closeExit animation is causing the wallpaper to
fade when finishing an activity. Since we are doing custom
animation when closing activities, we can replace the
windowAnimationStyle attribute 'activityCloseExitAnimation' with null
to make sure our animation is the only animation that is run when
ending an activity during setup.

This CL is responisible for activities in Settings during setup
only (wifi and device name).

Change-Id: Idd42a1cd3e4d9892cad926e87b8fc7e5871e619e