Fix custom name page bug upon interruption
Moved openInEditMode() from onViewCreated() to onResume().
Since this enter custom name IME session could be interrupted by
PairingActivity (or any other activity), we move openInEditMode() to
onResume() to make sure it is always in edit mode on resumed status.
It should be safe to invoke this method in onResume() as documentation
indicated that "it can be invoked in onViewCreated()", which happens
earlier than onResume().

b/38449791

Test: Manually

Change-Id: Iacb2e545da75e253242d6016cba526e29ea618fa