Alignment fix in Settings.
(1) Change layout.
(2) Change alignment behavior.

For the left part of the page, the keyline matches with the baseline of
the guidance title. If there are multiple lines, matches with the
baseline of second line, otherwise matches with the baseline of first
line.
For the right part of the page, the keyline matches with the mean line
of the title textview of the first item in the actions list.

Bug: b/37166263
Spec:
https://folio.googleplex.com/_/preview/0B2-gWRdLkQ5zMW9oS2c2VlBjNUU#%2F01.png%3Fz=width
Screenshot:
https://drive.google.com/open?id=0B6dTPljnkZXPczRzX3BNOGxjRUE

Test: Manual
Change-Id: I251d8bd8b63b6350cc6753c929282f36d204cd74