Fix location setting & NLP dialog interaction bug.
Due to the NLP dialog popup the location setting did not get the summary
update at all times. It's rather a corner case: only when user did not
previously agree to the NLP, the summary was not updated on location
state change. This bad behavior is fixed with this change.

Test: Manually verified the location setting's summary gets updated on
every location state change event.
Bug: 38027976
Change-Id: Ifba103b6681492da5fb7adf261db04c4cd09544d