Prompt PIN on boot if system user is FBE encrypted
This prevents an FBE device from getting stuck in an ugly locked state
after a reboot when adding a restricted profile.

Side effect: rebooting will now boot into the primary profile after
entering a PIN, instead of booting straight back into the restricted
profile. In this sense it appears to the user that part of rebooting is
resetting the current user back to the primary profile, which makes some
kind of intuitive sense.

Bug: 112088303
Test: make -j20 RunTvSettingsRoboTests
Change-Id: Icb20c27ac5cc1fda747cae79510b5f7ac65a6a0b