Check the current user's unlock state, not owner
This prevents a race condition this part of the check is supposed to
prevent in which the user has no lock screen but has not yet reached
unlocked state (ie. they show up as locked for a split second on an FBE
device before moving into unlocked).

Cherry-picked from commit 511307f03 with a small merge conflict in
FallbackHome.java

Bug: 113339509
Test: make -j20 RunTvSettingsRoboTests
Change-Id: I452dcef69fe46602d9e47c2ffeceb47cb8875b29