TvSettings: Fix handling of empty enabled surround formats setting
The code was using 'String.split' method which returns an array
with an empty string element when the input is an empty string.
This was causing 'parseInt' to throw. Whereas 'TextUtils.split'
returns an empty array for the empty string, and thus
'Arrays.stream' emits an empty stream, as expected.

Test: switch to manual surround formats selection, observe logcat
Change-Id: I37e02603bfeb0bc7ec913d631ef36c49234caf32