Switch users on LOCKED_BOOT_COMPLETED
Because rebooting the device effectively exists a restricted profile,
there is a BOOT_COMPLETED receiver which switches back into the profile
the device was in before reboot. Because of changes in Android P to when
BOOT_COMPLETED is sent, the receiver is too slow to switch users, which
gives a restricted user enough time to go through the settings of the
parent profile and change the allowed apps which defeats the purpose of
the restricted profile. To fix this, I changed the receiver so that it
instead switches users on LOCKED_BOOT_COMPLETED which is sent earlier.

Bug: 115609340
Test: make -j RunTvSettingsRoboTests
Change-Id: I53298b756bdc466bc61aa8829ceed1a91210a1a3