Set a NUMERIC PIN for TV restricted profiles
As of AOSP 843475 (bug 110172241) LockPatternUtils enforces correct
password quality to be used for passwords, by throwing an exception.

TvSettings was always wrong to use PASSWORD_QUALITY_SOMETHING as this
is only applicable to pattern lock types. Since the PIN is a 4 digit
number by default we squint a little bit and call it
PASSWORD_QUALITY_NUMERIC.

Change-Id: I469abaccd844a16e6dab6f1addd235f8f9d2863a
Fix: 130016699
Test: Manual flash and create profile. No longer throws exception.
Test: Automated test to follow.