Fix bugs for slice renderer and TwoPanelSettings.
(a) Fix the bug that when the user hit too quickly, Settings would
crash. If the previous animation is not done, post current back()
method.

(b) When the user hit back in a slice flow, the left side slice fragment
would move to the preview panel, at this point, as no lifecycle has
changed, the counter of the SliceLiveData for this slice fragment  would
not decrease itself. When the corresponding preference get focused again,
the slice fragment's SliceLiveData will not trigger onActive()(it
was already active), thus the slice fragment would not receive
onChanged(), it will forever show spinner in this case.

To make sure everything behave correctly:
(1) Make SettingsPreferenceFragment handle the behavior of
adding/removing focus listener. In this way, focus listener would be added
as soon as possible when the fragment is resumed, so when the slice is
updated and onChanged() is triggered in SliceFragment, it will always
have focus listener at hand.

(2) In onPreferenceFocused(), if the preview fragment we need is already in the preview panel,
we won't add it to fragmentmanager. To do this, check the preview
fragment class and its arguments.

(c) Clear code to solve several crash and known issues. Crash in
Usage&Diagnostics, animation issue for preview fragment after starting other activity
("keep activity" is disabled), make TvSettings guard against the
case slice.getHint() is null  when sliceprovider is not correctly set up.

b/133268205
b/133190119
b/132892859
b/130180526
b/133265951
b/132444683
b/133196834
b/140131879

Test: Manually test common cases animations, behavior after start
other activities, "Date&time" preference, "developer options",
accessibilty mode

Change-Id: I19d7380344085a946e776a1b5674200e4dcca37b