Ensure main panel fragment call onPause() when starting immersiveFragment.
ApplicationsStateCallbacks from settingslib in AppManagementFragment is tied to
lifecycle. If AppManagementFragment does not go through onPause()(when
starting the immersive fragment) and onResume()(when returning to the
AppManagementFragment), the callbacks cannot be correctly triggered.

Make sure the main panel fragment is removed in fragment
transaction when it is invisible so onPause() and onResume() is called
correctly.

Test: Manual
Change-Id: I3cd20c939ef7937ab6eb589dbb570f57603c6eb4