Fix showing Restricted Profile icon.
For the restricted user, RestrictedProfileActivityLauncherEntry is not
getting enabled which is the component that launcher checks in order
to display the icon or not. We need to trigger the user switch when
entering the restricted profile.

Bug: 144513725
Test: manual test
Change-Id: Ie44a13468704471ba5cda097c216f6daa7f51fa7
(cherry picked from commit 77ad0ebb1014e6ffa59e120c4bf0c4d10707200d)