Be able to login captive portal via Wifi settings
If captive portal is detected when connting to an AP,
open a web page of the portal. After successfully login,
close the page.

Modify the Fragment demostrating the wifi information.
If the user cannot browse the Internet by the
selected AP without limitation (e.g. the user do not
successfully login the captive portal), create
WifiConnectionActivity instead of showing the detailed
information.

Bug: 146088438
Test: The test contains two parts.
  - Open Wifi selection page and click an AP
  with captive portal functionality, then it should
  open a web page. Login by following the instruction,
  then Wifi should worked normally.
  - As above. Instead of login, press BACK button.
  It should return to the Wifi selection page.
  Click the same AP again and click 'Connect' in the
  following selections. A web page should be opened again.
  After login Wifi should worked normally as well.

Change-Id: Iced093b2be85ffbb2fd5cd3b1ec3ff38e15ca9ec