Add selectable state for slice preferences.
This is needed to correctly determine focus. If a preference is set to
disabled, it will still be selectable which means it is still focusable.
This is evident when the first preference option in a settings page is
disabled and it looks like there is no focus on the page.

Bug: 144431016
Test: Manual.
Change-Id: I1a23f67017432678c5520d999d100d7ee8adcf1b