[Slices API] Fix the bug that replaced slice fragment does not update.
This cl fix the following case:
a) Fragment B replace Fragment A which bind to the same slice
b) As no slice data is changed, Fragment B would not update.
The lines removed in this Cl was the workaround to fix some state
rebinding issue which looks like not a problem anymore.
Bug: 151179932

Test: Manual
Change-Id: I0b593c6537d34e5ae0c5e18f9eafc44ac0e390b2