Fix the crash in Manage keyboards for R branch.
A CL in settingslib(ag/8044701) is trying to set the size of icon by
giving it a new LayoutParams LinearLayout.LayoutParams,however, for TvSettings,
we are using FrameLayout in leanback_preference.xml which cause the crash.

I tried to branch the logic to handle two layout respectively.
However, setLayoutParams(new FrameLayout.LayoutParams(size, size))
renders a weird UI for our case - there is less space between title and
icon.

It turns out the simplest solution to fix this is setting the icon size
to be "0", thus the code setLayoutParams() would not be executed. UI
looks normal using this way.
And without touching the code in settingslib, we are guranteed to not affect
phone's UI.

Bug: 146654624

Test: Manually check one-panel and two-panel UI.
Change-Id: I876206ee2b54f541a0ec322e8144986d92a32146