Handle WIFI_ADD_NETWORKS by opening network settings
The CTS test
android.content.cts.AvailableIntentsTest#testAddNetworksIntent
checks that the WIFI_ADD_NETWORKS Intent is handled.

TvSettings does not yet handle that Intent.
As a first step, to pass CTS, open "Network & Internet" settings when
the Intent is sent.
In the future, TvSettings should list the network suggestions passed
via the Intent extras and allow the user to save the suggested networks.

Bug: 159609501
Bug: 159557103
Test: adb shell am start-activity -a android.settings.WIFI_ADD_NETWORKS
Change-Id: I9278a9344fdf90af1dc79e3e5d46e514d946b95e