Improve a11y for BT pairing screen - TvSettings
Normally, Talkback announces an Activity's label with it is visible and
announces components' descriptions as they are focused. The BT pairing
Activity is special as the user may not have a connected remote control
to focus on components and listen for the instructions.

This CL addresses this by concatenating the instructions and setting it
as the Activity's title to be announced by Talkback. (If the user has a
connected remote control, they can always start navigating and
interrupting the concatenated version of the instructions and listen to
them one by one.)

This CL also improves the accessibility of the extra instructions and
allows strings to be overrode to replace the remote button icons for
Talkback to announce.

Bug: 155090708
Test: When Talkback is on, it announces the full instruction when the
Activity is launched.

Change-Id: Ic305138a97771acc9c8f2369b4a727b174213b6a