Don't reset NumberPicker text when getting the PIN
We are at the end of an animation at this point, so the updated text
is going to be off by one until the next scroll event - which won't
happen because at this point we're about to reset all the fields.

It's correct to just leave the PIN UI alone and only get the values. The
UI will take care of itself next time something changes.

Test: Manual. Create profile. Change PIN. Enter. Exit. Delete.
Fix: 165939382
Change-Id: Ife3c18414ceb9cee528086711e0c7dd4fd2715fe