Wifi password keyboard actionNext to actionDone
This CL changes actionNext to actionDone for the EditText associated
with inputting a Wifi password in settings. Since the GuidedStepFragment
in the leanback library overrides the value of android:imeOptions in
setup_password_item.xml, I have removed it to avoid future confusion on
where the IME options are being set.

Bug: 164148849
Test: manual
Change-Id: I09c2c97e023ac5be44ef9378b99d9b7df3ab2129