Align TwoPanelSettings and TvSettings manifests
TwoPanelSettingsOverlay is missing a large number of fixes that were
made for R and S. Likewise, TvSettings is missing a large number of
TwoPanel-specific fixes.

This change reduces the delta between manifests to just the important
lines. Next step is to move the remaining items into more appropriate
overlay formats.

 * Enabled/disabld components should be replaced by component-overrides
 * Themes should be replacd by RRO

Test: make checkbuild; make TwoPanelSettingsOverlay; # and manual
Test: atest TvSettingsRoboTests
Bug: 174343446
Change-Id: I2abc062e7cebea8d045fb92e283c64805fe101ca