Allow a11y focus on InfoFragments
This works by making focusable and requesting focus for an otherwise
non-focusable InfoFragment IFF a11y is turned on. The previous view
is cached so that focus can be restored on backwards movement without
going through moveToPanel(). Furthermore, the releaseA11yFocus does not
check if talkback is on purposedly. This will allow correct behavior if,
on the off-chance, that a user turns a11y off while focused on an info
fragment. The intercepting behavior is in the back() method.

Please note that my preferred approach to this is for Talkback to handle
focus transitions, but it seems like the view/fragment hierarchy for
two panel settings makes this a bit convoluted and not just a
setImportantForAccessibility = yes. I will document more details in a doc.

Bug: 179693512
Test: Manually.

Change-Id: I1d1ee24c90a5a419567408269eaec1a1617acbd1