Improve initial a11yFocus on Panel Change
Set the previewPanel to be important for a11y as soon as possible
in the navigate to preview flow.

Additionally, put an animation finished listener on our slide animation
to request the focus. This should give the a11y screenreader enough time
to correctly build the traversal graph and know where the a11y focus
should be set.

The only drawback here is that forward navigation will still require the
user to wait 1-2s after the preview fragment loads when a new preference
has be selected/focused in the main fragment/panel in order to work.
This is likely because the accessibility traversal graph hasn't had
enough time to build after we replace the preview panel with the new
fragment. I am currently trying to find a way to force this to happen
faster.

Bug: 179693512
Test: Manual, with an actual remote and TV. Not vysor or adb keyevents
Change-Id: I297f62ee757277b043fee50a76e7bd63f0c606b4