Add a11y pane title behavior for one panel settings
This will give some context on fragment transitions by
announcing the titles. This doesn't handle focus or any
immersive fragments that may be in settings.

It is important to not allow this behavior in TwoPanel
settings since it's constantly creating new fragments so
we would be updating the title way too frequently, which
results in a mess of announcements.

Bug: 179693512
Test: Manual, on one and two panel settings.
Change-Id: I60a79d5e5b5971dc0c5a757a8b290cb78656883a