Allow InfoFragments focus in a11y mode
This allows focus to go to InfoFragments when a11y is on, but does not
scroll our panels, therefore the focus stays on the preview panels and
we do not have a fragment change. This allows us to go back and forth
and read context to the user.

The user can exit with either the Dpad or by pressing the back button,
but we must intercept the back button in order to not trigger a scroll
backwards and only set the focus on the main panel again.


Bug: 179693512
Test: manual
Change-Id: I954380064ac0bee634ba76004cf870410a9aea45