Make TV discoverable during manual pairing
Occasionally, another device initiates pairing to an Android TV. To
accommodate this, ATVs are discoverable by default, but discoverability
was turned off on Chromecast with Google TV. For externally-initiated
pairing to work on Chromecast, turn on discoverability when manual pairing
is enabled.

Bug: 181858303
Test: Click "Pair remote or accessory", then back out. Observe via logs
and dumpsys bluetooth_manager the correct scan mode before / during /
after.
Test: Pair an Android phone to the Chromecast. (Note: when the phone is
paired, the Settings app stays on the scanning screen, unlike when a
headset is paired. Requires pressing "pair" on both devices.)

Change-Id: I6cf659b350e4196b1d4f107c8464bada71dd8d62