Append error message to end of prefFrag content description
When the error message is sent to TwoPanelSettingsFragment, the
error message is appended to the title and sent to the prefence
to modify the local copy of the content description. After, the
item adapter is notified that an item is changed, this will allow
the recyclerview to rebind the viewholder, which is needed to
actually set the content description on the view. Last, the onBind
now sets the itemView to be a polite live region when there is
a custom content description so that any live modification/update
of an error is announced immediately instead of waiting for the
focus to leave and return to trigger another annoucement.

Bug: 185481056
Test: Manually (screenrecord posted to bug)
Change-Id: I18b73840414ef270b71b446f97290115d534a611