Remove System Update Intent started from preference xml
The system update activity was started twice in some cases:
Once in code from the AboutFragment on preference click,
and once from the preference itself if the intent tag was present in the
preference xml and no fragment was set on the preference.

This change removes the intent tag in the preference xml for all flavors
and only starts the system update activity from the AboutFragment.

Bug: 172463049
Test: Manually, with TvSettings in different flavors
Change-Id: I7ec4dbdebf49969729f25f2cf8e74f9315ae1534