Use restricted preference for Location status if location is restricted
If the user restriction DISALLOW_SHARE_LOCATION or
DISALLOW_CONFIG_LOCATION is set by an admin, then show a restricted
preference instead of the regular ListPreference.
This visually indicates that the preference is restricted, hides
the list options in the right panel in the two panel layout,
and shows an explanatory message dialog on click.

Bug: 181201913
Test: CtsVerifier
Change-Id: Ic581b72fc4a80d22da57f3dde7b1d3e439b682f6