Check user restrictions in date & time settings
If the DISALLOW_CONFIG_DATE_TIME user restriction is set,
these preferences will be restricted:
- Automatic date & time
- Set date
- Set time
- Set time zone

This preference remains unrestricted:
- Use 24-hour format

A RestrictedPreferenceAdapter is introduced that replaces the original
preference with a RestrictedPreference if the user restriction is set
by an admin. This leads to a consistent user experience and works for
any type of preference.

Bug: 181201913
Test: CtsVerifier
Change-Id: Ib91f1637dc8d3c8633ac8a683618bac13283058e