Hide unsaved networks if DISALLOW_CONFIG_WIFI is set
CtsVerifier tests that if the DISALLOW_CONFIG_WIFI user restriction is
set, then you cannot view WiFi networks in range.
This change marks WiFi networks invisible if the restriction is set and
the wifi hasn't been previously saved.
The "See more/less" item is updated to only show if there are enough
*visible* preferences in the category.

Bug: 181201913
Test: CtsVerifier
Tested-On-1Panel: On adt3, wifi settings look unchanged
Tested-On-2Panel: On adt3_gtv, with & without restriction set
Change-Id: I63851930e9bf1d70e3eddb2be8da9742da012a06