Prevent modifying of networks set up by device owner if locked down
If Settings.Global.WIFI_DEVICE_OWNER_CONFIGS_LOCKDOWN is set,
wifi networks created by the device/profile owner should not be
modifiable.
This change restricts the "Proxy settings", "IP settings", and "Forget
network" preferences in the wifi details fragment.

Also, when the EditIpSettingsActivity or EditProxySettingsActivity is
started, an informational dialog is shown and the activity finishes
immediately.

Finally, the AdvancedOptionsState, ProxySettingsState, and
IpSettingsState states of the AdvancedWifiOptionsFlow will skip to the
end of the advanced wifi options flow when navigated to.

Bug: 181201913
Test: CtsVerifier -> Device Owner Tests -> WiFi configuration lockdown
Test-On-1Panel: On adt3, unrestricted, no change
Test-On-2Panel: On adt3_gtv, with and without restriction
Change-Id: I629252bb90ae9fa39bddca24ab6f5a7fce2db9e4