Check DISALLOW_CONFIG_WIFI when starting wifi configuraton activities
When a wifi configuration activity is started, check if the user has the
DISALLOW_CONFIG_WIFI restriction set.
If it is set by an admin, show an explanatory dialog.
If the restriction is set, finish the activity immediately.

Bug: 181201913
Test: CtsVerifier -> Device Owner Tests -> Disallow configuring WiFi
      Set restriction
Test: adb shell am start -a com.android.settings.wifi.action.EDIT_IP_SETTINGS --ei network_id 1
Test: adb shell am start -a com.android.settings.wifi.action.EDIT_PROXY_SETTINGS --ei network_id 1
Test: adb shell am start com.android.tv.settings/.connectivity.setup.WifiSetupActivity
Test: adb shell am start com.android.tv.settings/.connectivity.WifiConnectionActivity
Test: adb shell am start com.android.tv.settings/.connectivity.AddWifiNetworkActivity
Test-On-1Panel: On adt3, unrestricted, no visible change
Test-On-2Panel: On adt3_gtv, with restriction set, activities finish and
                show dialog
Change-Id: I93c3ea367c1f57d18c7ce0ebdc32717edab29cf2