Add "Alarms & Reminders" special app access page
This screen lets users toggle which apps are allowed the
OP_SCHEDULE_EXACT_ALARM appop.
Also accessible via the intent
android.settings.REQUEST_SCHEDULE_EXACT_ALARM.
If a data URI with package scheme is passed, it only opens the page
showing the list of apps.

Bug: 181284165
Test: Navigate: Apps > Special app access > Alarms & reminders
Test: Navigate: Privacy > Special app access > Alarms & reminders (GTV)
Test: Toggle switch for Shell and check effect with `dumpsys appops`
Test: adb shell am start -a android.settings.REQUEST_SCHEDULE_EXACT_ALARM
Test: adb shell am start -a android.settings.REQUEST_SCHEDULE_EXACT_ALARM -d package://com.android.shell
Tested-On-1Panel: On adt3
Tested-On-2Panel: On adt3_gtv
Change-Id: I7bd6d16f102dd9d5bc664e394724ccecc52cd7b0