Fix adding and removing of preference focussed listeners
When the container Activity's onPause is invoked, it is possible that
the preference focussed listeners are removed while user contains
scrolling through the list. The corresponding preview fragment thus does
not get shown. In order to safeguard against this, always add listeners
when a fragment becomes the main panel, add them after the main panel's
view is created in onViewCreated and remove them when the fragment view
is destroyed.

Bug: b/191214995
Test: Ran throught multiple runs of app
Change-Id: I1a76d88d7edd316c815dcb7842a4f01fec06c861