Revert "Revert "Fix adding and removing of preference focussed listeners""
This reverts commit cf44548dca37a962f1df74b1a01f987cb598e31f.

Reason for revert: Fix issue in when SliceFragment is shown. On navigating back from SliceFragment, the listeners are correctly attached. The issue was due to attaching listeners in onCreateView. The attaching failed since the view was not created and thereby there was no VerticalGridView to attach the listeners to. Moved this to onViewCreated at which it is certain that the child fragment view has been inflated.

Bug: b/191214995
Test: Tested on 2 panel, went over all settings menus, in a two pass
manner, navigate into menu and then back and check scroll

Change-Id: I94007c7c7a712e6fae146e11a988ae669bcdc3cc