Make a11y workaround more specific in HelpFragment
Previously, we only checked if *any* a11y service was turned on, which
may end up breaking the send feedback button click since not all a11y
services will have an accessibility focus. This CL makes the check more
specific to be the TalkBackService that is enabled, therefore
guaranteeing that we have an accessibility focus feature currently
turned on.

Test: Manually with adt3-gtv
Bug: 174776406
Bug: 165694549

Change-Id: Ic15b4edc9b210f09c4c1a44dcf9bf425c4bb556b