Prevent overlapping of immersive and preview fragments
OnPreferenceTreeClick can be invoked before a runnable to show the
preview panel is queued. If no preview panel has been shown for the
current main panel, the fragment part of this preference being clicked
gets opened as an immersive fragment. After a delay, the preview panel
is also shown and this results in the overlap. To fix this, only treat
fragments that are instances of GuidedStepSupportFragment as started in
immersive mode.

Bug: b/192329938, b/163486240
Test: Verified on 2 panel settings

Change-Id: I17679a9b7889880374ad8ad574b2b8bce16e1d61