Add support embedded slice pref in slice
Embedded slice preference is used so that the slice provider
can dynamically control the visibility, title, summary and icon
of the placeholder preference, making the footprint in TvSettings
truly minimal. The embedded slice pref is used both as a
placeholder and a shadow preference for carrying the meta data
to be copied over. When used as a shadow preference it should
never be visible.

Current AOSP implementation only supports it when used in the
XML and not when specified in a slice. This means that the only
way the embedded slice is constructed is when the XML is inflated
and the route of constructing the embedded slice via the unpack
code in SlicePreferenceUtil.getPreference is unused and untested
(has bug). Via the XML route the shadow embedded
slice pref is never added into the view tree.

This implementation adds support for embedding a slice placeholder
via a slice, where the properties dynamically can be controlled
from another slice. To avoid the shadow embedded slice pref to
turn up in the view tree and for better code design a new type
is added. Also a new method in the builder is added for this.

It also fixes a null pointer exception in the previously unused
EmbeddedSlicePreference constructor.

Bug: 211421222
Test: Create a slice provider adding a second slice as embedded
      using the addEmbeddedPreference builder API and from the
      second slice using setEmbeddedPreference to control
      visibility, title, summary.
Change-Id: I46806b760d6df2c6fff5a75a088fc90666077761