Fix NPE after slice-pref target uri change
The Slice implementation handles multiple different Preferences
depending on what properties that are set by the slice provider
in the row builder.
In this case, the set properties (toggling between
slicetarget=available and slicetarget=null) causes the SliceFragment
to, even though the preference was using the same key, consider them
to be the same but not store the newly added slicetarget since it had
previously been removed. This causes a NullPointerException if not
handled. Solution is that, even though the old item could be re-used
by just transferring the new data to it, the old preference is
removed and replaced by the new one.

Bug: 211470801
Test: Step in and out of a slice preview where the slice target
is set and unset. Confirm no crash.

Change-Id: Ib1c52dfbfa5421dba66a688abe0a345fba022398