Add end icon support for Slice preference
Added support for end icon to the Slice preference API.
This means that it is now possible to set an icon to the right of a
slice preference item (left in RTL layout) in same way as you would
set an icon to the left. Only one of these icons can be set at the
time.

Bug: 211474361
Test: Set an icon with PreferenceSliceBuilder.setEndIcon
Example:
sliceBuilder.addPreference(new PreferenceSliceBuilder.RowBuilder(uri)
.setKey("key_end_icon_demo")
.setTitle("title")
.setEndIcon(IconCompat.createWithResource(context, R.drawable.endicon))
.setContentDescription("accessibility string for the end icon")
.setTargetSliceUri("content://com.test.myapp.slice/underlaying_slice"));

Change-Id: I48d403fc4a2844f944035e031831bd64d5aba252