Info Panel for physical privacy enabled info
On the camera/microphone privacy pages a warning is shown if the sensor
is disabled by a physical hardware switch.
With this change, this warning is selectable, and shows an info panel
when selected (for 2-Panel settings) / when clicked (for 1-Panel
settings).
The text in the info panel is customizable by overlaying
- camera_physical_privacy_enabled_panel_image
- camera_physical_privacy_enabled_panel_text
- microphone_physical_privacy_enabled_panel_image
- microphone_physical_privacy_enabled_panel_text

The 1-Panel unblock info page is dismissed automatically when the sensor
becomes unblocked.

Bug: 220716604
Test: manual on device with mic switch
Test-On-1Panel: manual
Test-On-2Panel: manual
Change-Id: I371951e9bab5f7a13b0b007af468edc4a68df36f