Parse provided ordered preference list
The ordered preference list meta data is recursively parsed to ensure
all the preferences and nested preferences in preference groups are
added. Each preference is represented by its key. This list represents
the intended final ordering of how the Preferences have to be laid out
on the screen.

Bug: 230894823
Test: This change in and of itself is a no-op. The overall customization
has been tested.

Change-Id: I703a7b1af399d9c5c92acbe4a9328329391de596