Fix bug where accessibility focus erroneously changes itself.
The method resetA11yFocusIfNeeded resets the accessibility focus after 1 s. In the meantime the user may have navigated to a different View, in particular one that is not input-focusable. Then the focus would be reverted which is not intended. Note that with the new TalkBack version the focus should never be lost and is in the worst case restored at least every 3 s.

Bug: 259220340
Test: Tested locally on ADT-3 GTV, actually on branch sc-tv-dev. Only tested with screen "Remotes & Accessoires".
Change-Id: I742871f1af02685242a249a798e1085ab01a99e3