TvSettings: resolution switch[1/1]
PD#SWPL-97686
BUG=254401735

Problem:
The presentation and and resolution of the dialog is displayed asynchronously,
and switching the resolution will cause the activity lifecycle to be reset.
Therefore, one case is likely to happen: the dialog will lose the dependent
window when the activity is destroyed and a dismiss is performed.

Solution:
1, destroy the CountDownTimer when the selection is made with the prompt of the dialog box.
2, add a condition to determine the existence of the activity attached to the dialog before "dialog.dismiss".

Verify:
local

Change-Id: Ifa1ce6502288e92a1f4192a2756b3c65a31f97a7
Signed-off-by: Lele Liu <lele.liu@amlogic.com>