Add config to disable Energy Mode selection UI
If Low Power Standby is not supported on a device, no related settings
will be shown.
If the config is set to false, TvSettings will show a simple toggle to
enable/disable Low Power Standby.
If the config is set to true, TvSettings will show a list of Energy
Modes to select from.

Bug: 277726800
Test: manual
Change-Id: If3aa2de0a1577177226bd4d4c910a67e66f215f9