Add energy mode color, base icons
Setup will show an energy mode selection screen with energy modes
represented with a color and icon.
The icon defined for the energy modes currently is the preference icon
with the default compound background baked in.
To provide Setup with a clean base icon, the compound icon overlay is
removed, and instead the preference icon is put together in code from
the default background and the energy mode icon.

Bug: 277889148
Test: manual
Change-Id: I64f98131e7dd2f994a2a77fcc9436a4cacb21c73