TvSettings: resolution switch[1/1]
PD#SWPL-119999
BUG=278231538

Problem:
When switching between different screen resolutions with different pixel densities,
the current activity is recreated. This can cause the "isActivityEnable" method to
receive an empty activity object, which does not meet the condition for "dialog.dismiss",
resulting in the dialog not being successfully destroyed.

Solution:
By removing the "isActivityEnable" method and only keeping the logic to destroy
the CountDownTimer after the dialog executes with "ok" or "cancel",
the "dialog.dismiss" method can be successfully executed.

Verify:
androidT-ADT4

Change-Id: Ie998770653ea4b56c730e32ec625eb3b1703e356
Signed-off-by: junlu.jiang <junlu.jiang@amlogic.com>