Use platform default for attentive timeout, don't set if pref is hidden
The attentive timeout value determines when the device goes to sleep
even if wakelocks are held. TvSettings provides several options to
choose from (4h / 6h / 8h / never).
When the EnergySaverFragment is created, it reads the current attentive
timeout and sets the preference's selection to the closest matching
option TvSettings provides.

However, it used a wrong default if the setting has not been set yet:
`DEFAULT_SLEEP_TIME_MS` = 20 minutes. (This also looks to be a typo,
as it was likely meant to use `DEFAULT_ATTENTIVE_TIME_MS`).
The "closest" option that TvSettings provides is 4h, so whenever the
Fragment is first created, since the timeout has not been set before, it
will be set to 4 hours.
The platform's PowerManager however doesn't use a fixed 4 hours default,
but reads its default from `config_attentiveTimeout`.
If `config_attentiveTimeout` is set to eg. `-1`, the attentive timeout
would be disabled initially, but once the Fragment is loaded, the
setting would be incorrectly set to 4 hours without the user actually
choosing a different timeout setting.

The attentive timeout preference is not shown by default in TvSettings
either (requires setting `config_show_standby_timeout`), so users
couldn't change the incorrectly set timeout value.

This change attempts to mitigate this issue, by resetting the attentive
timeout back to its platform default value on devices that do not give
the option to change the attentive timeout in TvSettings.

Bug: 286356445
Test: Manual, with setting shown/hidden, different timeout defaults
Change-Id: I02bf23018b7e8a6bd1e787d433e1934de68f10d1