Reduce calls to TransitionManager.delayedAnimation
Preferences (devices) added to the fragment may not be
attached to the correct parent during the animation. This
may cause two fragments to be displayed.

beginDelayedAnimation is now only called when the preference
list changes from empty to non-empty and vice versa. Preference
is also cleared in onStart to avoid animation jank on resume.

Bug: 259481424
Bug: 278622358
Bug: 291726039
Test: navigate quickly between Settings / Devices and Home
Change-Id: I70fa95d45328db79772ed89a411eb09e3a39810d